requirements.txt
flask 
flask-login
flask_sqlalchemy
sqlalchemy
python-decouple
